fix(codex): close activity re-subscribe leak, remove dead snapshot boundary, correct docs
Codex round-1 review of the EF1 activity cutover. Dispositions:

- F2 (major, fixed): the padi kaval re-subscribe delay added an abort listener
  every retry and never removed it on timeout — an unbounded listener leak on the
  subscription-lifetime signal across a sustained kaval outage. Extracted the
  existing correct pattern to `padi/src/abortableDelay.ts` (removes on both exits,
  timer unref'd) and reused it in both `liveActivity.ts` and `inventoryReconcile.ts`
  (deleting the latter's identical module-local `delay`), one source of truth.
- F5 (nit, fixed): `createSnapshotBoundary` in Terminal.tsx was created + armed but
  its only observable (`isLiveDelta()`) lost its sole caller with the attach-sink
  lighting — dead. Removed it and corrected the comment; the whole `xterm-kit`
  snapshotBoundary module + test + exports were then fully orphaned, so deleted.
- F6 (nit, fixed): the `useTerminalActivity` comment claimed the stream "declares
  its own error policy on the spec" — StreamSpec has no such field. Rewrote it: the
  bare `.use()` rides the framework's default STREAM_RETRY reconnect/health.
- F4 (minor, fixed): the canonical architecture doc's data-flow section omitted the
  new kaval-owned activity flow. Updated prose + SVG title/label per architecture.md.
- F1 (major, disputed): the pollOnEvent read→install gap can't lose an edge — edges
  only arrive via the kaval subscription, which isn't established in that microtask
  window; the channel has no replay; install is once-per-subscription. Not reachable.
- F3 (minor, disputed): the sticky-live fence (incl. live→empty) is unit-tested via
  activityFrameDiff; the proxy-as-prev footgun is a single-caller snapshot discipline
  (lens hickey-3 agreed no change needed).
